+++ doc.go
@@ -38,7 +38,7 @@ Source code and other details for the project are available at GitHub:
 
 # Current Version
 
-0.1.8
+0.1.11
 
 # Installation
 
@@ -266,24 +266,23 @@ Outline:
 	//
 	// For example, if you want to print using a logrus
 	// logger you can do the following:
-	// `golog.Install(logrus.StandardLogger())`
 	//
-	// Look `golog#Handle` for more.
-	Install(logger ExternalLogger)
-
-	// InstallStd receives  a standard logger
-	// and automatically adapts its print functions.
+	//	Install(logrus.StandardLogger())
 	//
-	// Install adds a golog handler to support third-party integrations,
-	// it can be used only once per `golog#Logger` instance.
+	// Or the standard log's Logger:
 	//
-	// Example Code:
 	//	import "log"
 	//	myLogger := log.New(os.Stdout, "", 0)
-	//	InstallStd(myLogger)
+	//	Install(myLogger)
 	//
-	// Look `golog#Handle` for more.
-	InstallStd(logger StdLogger)
+	// Or even the slog/log's Logger:
+	//
+	//	import "log/slog"
+	//	myLogger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
+	//	Install(myLogger) OR Install(slog.Default())
+	//
+	// Look `golog#Logger.Handle` for more.
+	Install(logger ExternalLogger)
 
 # Logrus Integration
 
@@ -350,7 +349,7 @@ Example Code:
 
 	func main() {
 		golog.SetLevel("error")
-		golog.InstallStd(myLogger)
+		golog.Install(myLogger)
 
 		golog.Debug(`this debug message will not be shown,
 		because the golog level is ErrorLevel`)
@@ -372,4 +371,4 @@ Examples:
 package golog
 
 // Version is the version string representation of the "golog" package.
-const Version = "0.1.8"
+const Version = "0.1.11"